+++ CHANGELOG.md
@@ -10,6 +10,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.32.29] - 2026-07-11
+
+### Changed
+
+- Maintenance: sync ref:GH#27144 to TODO.md [skip ci]
+- Maintenance: sync ref:GH#27139 to TODO.md [skip ci]
+
+### Fixed
+
+- defer canonical lint policy migration (#27142)
+
+## [3.32.28] - 2026-07-11
+
+### Changed
+
+- feat: provision repository lint policy (#27136)
+
 ## [3.32.27] - 2026-07-11
 
 ### Changed